CRUD terminé

+++ controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
    const bookObject = JSON.parse(req.body.book);
@@ -22,7 +23,7 @@ exports.modifyBook = (req, res, next) => {
     } : { ...req.body };
 
     delete bookObject._userId;
-    book.findOne({_id: req.params.id})
+    Book.findOne({_id: req.params.id})
     .then((book) => {
         if (book.userId != req.auth.userId) {
             res.status(400).json({ message : 'Non-autorisé' });
@@ -39,9 +40,22 @@ exports.modifyBook = (req, res, next) => {
 }
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(400).json({ error }))
+    Book.findOne({ _id: req.params.id})
+    .then(book => {
+        if (book.userId != req.auth.userId) {
+            res.status(401).json({message: 'Non-autorisé'});
+        } else {
+            const filename = book.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Book.deleteOne({_id: req.params.id})
+                .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })  
+    .catch( error => {
+        res.status(500).json({ error });
+    });
 }
